+++ common/src/main/java/com/gregtechceu/gtceu/api/data/worldgen/WorldGeneratorUtils.java
@@ -1,8 +1,9 @@
-package com.gregtechceu.gtceu.api.data.worldgen.generator;
+package com.gregtechceu.gtceu.api.data.worldgen;
 
 import com.google.common.collect.HashBiMap;
 import com.gregtechceu.gtceu.api.data.worldgen.GTOreFeatureEntry;
 import com.gregtechceu.gtceu.api.data.worldgen.IWorldGenLayer;
+import com.gregtechceu.gtceu.api.data.worldgen.generator.VeinGenerator;
 import com.gregtechceu.gtceu.api.registry.GTRegistries;
 import com.gregtechceu.gtceu.data.recipe.CustomTags;
 import com.mojang.serialization.Codec;
@@ -25,8 +26,8 @@ public class WorldGeneratorUtils {
     private static final Map<WorldGenLevel, WorldOreVeinCache> oreVeinCache = new WeakHashMap<>();
 
     public static final Map<String, IWorldGenLayer> WORLD_GEN_LAYERS = new HashMap<>();
-    public static final HashBiMap<ResourceLocation, Codec<? extends GTOreFeatureEntry.VeinGenerator>> VEIN_GENERATORS = HashBiMap.create();
-    public static final HashBiMap<ResourceLocation, Function<GTOreFeatureEntry, ? extends GTOreFeatureEntry.VeinGenerator>> VEIN_GENERATOR_FUNCTIONS = HashBiMap.create();
+    public static final HashBiMap<ResourceLocation, Codec<? extends VeinGenerator>> VEIN_GENERATORS = HashBiMap.create();
+    public static final HashBiMap<ResourceLocation, Function<GTOreFeatureEntry, ? extends VeinGenerator>> VEIN_GENERATOR_FUNCTIONS = HashBiMap.create();
 
 
     private static class WorldOreVeinCache {
