+++ common/src/main/java/com/gregtechceu/gtceu/common/data/GTFeatures.java
@@ -2,7 +2,7 @@ package com.gregtechceu.gtceu.common.data;
 
 
 import com.gregtechceu.gtceu.GTCEu;
-import com.gregtechceu.gtceu.api.data.worldgen.generator.*;
+import com.gregtechceu.gtceu.api.data.worldgen.modifier.*;
 import com.gregtechceu.gtceu.api.data.worldgen.GTOreFeature;
 import com.gregtechceu.gtceu.api.data.worldgen.GTOreFeatureConfiguration;
 import com.gregtechceu.gtceu.api.registry.GTRegistries;
