+++ common/src/main/java/com/gregtechceu/gtceu/core/mixins/PlacedFeatureMixin.java
@@ -2,7 +2,7 @@ package com.gregtechceu.gtceu.core.mixins;
 
 import com.gregtechceu.gtceu.api.data.worldgen.GTOreFeatureConfiguration;
 import com.gregtechceu.gtceu.api.data.worldgen.GTOreFeatureEntry;
-import com.gregtechceu.gtceu.api.data.worldgen.generator.VeinCountFilter;
+import com.gregtechceu.gtceu.api.data.worldgen.modifier.VeinCountFilter;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Holder;
 import net.minecraft.util.RandomSource;
