+++ common/src/main/java/com/gregtechceu/gtceu/integration/kjs/GregTechKubeJSPlugin.java
@@ -14,6 +14,7 @@ import com.gregtechceu.gtceu.api.data.chemical.material.stack.MaterialStack;
 import com.gregtechceu.gtceu.api.data.chemical.material.stack.UnificationEntry;
 import com.gregtechceu.gtceu.api.data.tag.TagPrefix;
 import com.gregtechceu.gtceu.api.data.worldgen.*;
+import com.gregtechceu.gtceu.api.data.worldgen.generator.VeinGenerator;
 import com.gregtechceu.gtceu.api.gui.GuiTextures;
 import com.gregtechceu.gtceu.api.item.tool.GTToolType;
 import com.gregtechceu.gtceu.api.machine.MachineDefinition;
@@ -236,9 +237,9 @@ public class GregTechKubeJSPlugin extends KubeJSPlugin {
                     .flatMap(DataResult::result)
                     .orElse(null);
         });
-        typeWrappers.register(GTOreFeatureEntry.VeinGenerator.class, (ctx, o) -> {
+        typeWrappers.register(VeinGenerator.class, (ctx, o) -> {
             return Optional.ofNullable(NBTUtils.toTagCompound(o))
-                    .map(tag -> GTOreFeatureEntry.VeinGenerator.DIRECT_CODEC.parse(NbtOps.INSTANCE, tag))
+                    .map(tag -> VeinGenerator.DIRECT_CODEC.parse(NbtOps.INSTANCE, tag))
                     .flatMap(DataResult::result)
                     .orElse(null);
         });
